# openapi.yml
openapi: 3.0.3
info:
  title: OpenAI API Mock Server
  description: Mock server compatible with OpenAI API
  version: 1.0.0
servers:
  - url: /v1
paths:
...
  /models/{model}:
    get:
      operationId: retrieveModel
      summary: Retrieve a model
      description: Retrieves a model instance.
      tags:
        - Models
      parameters:
        - name: model
          in: path
          required: true
          schema:
            type: string
          description: The ID of the model to use for this request.
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Model'
  /chat/completions:
    post:
      operationId: createChatCompletion
      summary: Create chat completion
      description: Creates a model response for the given chat conversation.
      tags:
        - Chat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChatCompletionRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateChatCompletionResponse'
  /completions:
    post:
      operationId: createCompletion
      summary: Create completion
      description: Creates a completion for the provided prompt and parameters.
      tags:
        - Completions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCompletionRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateCompletionResponse'
  /embeddings:
    post:
      operationId: createEmbedding
      summary: Create embeddings
      description: Creates an embedding vector representing the input text.
      tags:
        - Embeddings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEmbeddingRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateEmbeddingResponse'
  /responses:
    post:
      operationId: createResponse
      summary: Create a model response
      description: Creates a model response using the Responses API.
      tags:
        - Responses
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateResponseRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateResponseResponse'
components:
  schemas:
...
    Model:
      type: object
      required:
        - id
        - object
        - created
        - owned_by
      properties:
        id:
          type: string
          description: The model identifier.
        object:
          type: string
          enum: [model]
        created:
          type: integer
          format: int64
          description: Unix timestamp of when the model was created.
        owned_by:
          type: string
          description: Organization that owns the model.
    CreateChatCompletionRequest:
      type: object
      required:
        - model
        - messages
      properties:
        model:
          type: string
          description: ID of the model to use.
        messages:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/ChatCompletionRequestMessage'
        temperature:
          type: number
          minimum: 0
          maximum: 2
          default: 1
          description: Sampling temperature.
        top_p:
          type: number
          minimum: 0
          maximum: 1
          default: 1
          description: Nucleus sampling parameter.
        n:
          type: integer
          minimum: 1
          maximum: 128
          default: 1
          description: Number of completions to generate.
        stream:
          type: boolean
          default: false
          description: Whether to stream partial progress.
        stop:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
              minItems: 1
              maxItems: 4
          description: Stop sequences.
        max_tokens:
          type: integer
          description: Maximum tokens to generate.
        max_completion_tokens:
          type: integer
          description: Maximum completion tokens to generate.
        presence_penalty:
          type: number
          minimum: -2
          maximum: 2
          default: 0
        frequency_penalty:
          type: number
          minimum: -2
          maximum: 2
          default: 0
        logit_bias:
          type: object
          additionalProperties:
            type: integer
        user:
          type: string
          description: A unique identifier representing your end-user.
        seed:
          type: integer
          description: Seed for deterministic sampling.
        tools:
          type: array
          items:
            $ref: '#/components/schemas/ChatCompletionTool'
          description: A list of tools the model may call.
        response_format:
          $ref: '#/components/schemas/ChatCompletionResponseFormat'
    ChatCompletionTool:
      type: object
      required:
        - type
        - function
      properties:
        type:
          type: string
          enum: [function]
        function:
          $ref: '#/components/schemas/ChatCompletionToolFunction'
    ChatCompletionToolFunction:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        description:
          type: string
        parameters: {}
    ChatCompletionResponseFormat:
      type: object
      required:
        - type
      properties:
        type:
          type: string
          enum: [text, json_object, json_schema]
        json_schema:
          $ref: '#/components/schemas/ChatCompletionJSONSchemaSpec'
    ChatCompletionJSONSchemaSpec:
      type: object
      required:
        - name
      properties:
        name:
          type: string
        strict:
          type: boolean
        schema: {}
    ChatCompletionRequestMessage:
      type: object
      required:
        - role
        - content
      properties:
        role:
          type: string
          enum: [system, user, assistant, tool, function]
        content:
          type: string
        name:
          type: string
        tool_calls:
          type: array
          items:
            $ref: '#/components/schemas/ChatCompletionMessageToolCall'
        tool_call_id:
          type: string
        function_call:
          $ref: '#/components/schemas/ChatCompletionRequestMessageFunctionCall'
    ChatCompletionRequestMessageFunctionCall:
      type: object
      required:
        - name
        - arguments
      properties:
        name:
          type: string
        arguments:
          type: string
    ChatCompletionMessageToolCall:
      type: object
      required:
        - id
        - type
        - function
      properties:
        id:
          type: string
        type:
          type: string
          enum: [function]
        function:
          type: object
          required:
            - name
            - arguments
          properties:
            name:
              type: string
            arguments:
              type: string
    CreateChatCompletionResponse:
      type: object
      required:
        - id
        - object
        - created
        - model
        - choices
      properties:
        id:
          type: string
        object:
          type: string
          enum: [chat.completion]
        created:
          type: integer
          format: int64
        model:
          type: string
        choices:
          type: array
          items:
            $ref: '#/components/schemas/ChatCompletionChoice'
        usage:
          $ref: '#/components/schemas/CompletionUsage'
        system_fingerprint:
          type: string
        service_tier:
          type: string
    ChatCompletionChoice:
      type: object
      required:
        - index
        - message
        - finish_reason
      properties:
        index:
          type: integer
        message:
          $ref: '#/components/schemas/ChatCompletionResponseMessage'
        finish_reason:
          type: string
          enum: [stop, length, tool_calls, content_filter, function_call]
        logprobs:
          $ref: '#/components/schemas/ChatCompletionChoiceLogprobs'
    ChatCompletionResponseMessage:
      type: object
      required:
        - role
        - content
      properties:
        role:
          type: string
          enum: [assistant]
        content:
          type: string
          nullable: true
        refusal:
          type: string
          nullable: true
        tool_calls:
          type: array
          items:
            $ref: '#/components/schemas/ChatCompletionMessageToolCall'
        function_call:
          $ref: '#/components/schemas/ChatCompletionResponseMessageFunctionCall'
...
    ChatCompletionChoiceLogprobs:
      type: object
      nullable: true
      properties:
        content:
          type: array
          nullable: true
          items:
            $ref: '#/components/schemas/ChatCompletionTokenLogprob'
...
    CompletionUsage:
      type: object
      required:
        - prompt_tokens
        - completion_tokens
        - total_tokens
      properties:
        prompt_tokens:
          type: integer
        completion_tokens:
          type: integer
        total_tokens:
          type: integer
        prompt_tokens_details:
          $ref: '#/components/schemas/PromptTokensDetails'
        completion_tokens_details:
          $ref: '#/components/schemas/CompletionTokensDetails'
    PromptTokensDetails:
      type: object
      properties:
        cached_tokens:
          type: integer
        audio_tokens:
          type: integer
    CompletionTokensDetails:
      type: object
      properties:
        reasoning_tokens:
          type: integer
        audio_tokens:
          type: integer
        accepted_prediction_tokens:
          type: integer
        rejected_prediction_tokens:
          type: integer
    CreateCompletionRequest:
      type: object
      required:
        - model
        - prompt
      properties:
        model:
          type: string
        prompt:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
        suffix:
          type: string
          nullable: true
        max_tokens:
          type: integer
          minimum: 0
          default: 16
        temperature:
          type: number
          minimum: 0
          maximum: 2
          default: 1
        top_p:
          type: number
          minimum: 0
          maximum: 1
          default: 1
        n:
          type: integer
          minimum: 1
          maximum: 128
          default: 1
        stream:
          type: boolean
          default: false
        logprobs:
          type: integer
          minimum: 0
          maximum: 5
          nullable: true
        echo:
          type: boolean
          default: false
        stop:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
              minItems: 1
              maxItems: 4
        presence_penalty:
          type: number
          minimum: -2
          maximum: 2
          default: 0
        frequency_penalty:
          type: number
          minimum: -2
          maximum: 2
          default: 0
        best_of:
          type: integer
          minimum: 0
          maximum: 20
          default: 1
        logit_bias:
          type: object
          additionalProperties:
            type: integer
        user:
          type: string
        seed:
          type: integer
    CreateCompletionResponse:
      type: object
      required:
        - id
        - object
        - created
        - model
        - choices
      properties:
        id:
          type: string
        object:
          type: string
          enum: [text_completion]
        created:
          type: integer
          format: int64
        model:
          type: string
        choices:
          type: array
          items:
            $ref: '#/components/schemas/CompletionChoice'
        usage:
          $ref: '#/components/schemas/CompletionUsage'
        system_fingerprint:
          type: string
    CompletionChoice:
      type: object
      required:
        - index
        - text
        - finish_reason
      properties:
        index:
          type: integer
        text:
          type: string
        logprobs:
          $ref: '#/components/schemas/CompletionChoiceLogprobs'
        finish_reason:
          type: string
          enum: [stop, length, content_filter]
...
    CreateEmbeddingRequest:
      type: object
      required:
        - model
        - input
      properties:
        model:
          type: string
        input:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
        encoding_format:
          type: string
          enum: [float]
          default: float
        dimensions:
          type: integer
          minimum: 1
        user:
          type: string
    CreateEmbeddingResponse:
      type: object
      required:
        - object
        - data
        - model
        - usage
      properties:
        object:
          type: string
          enum: [list]
        data:
          type: array
          items:
            $ref: '#/components/schemas/Embedding'
        model:
          type: string
        usage:
          $ref: '#/components/schemas/EmbeddingUsage'
    Embedding:
      type: object
      required:
        - index
        - object
        - embedding
      properties:
        index:
          type: integer
        object:
          type: string
          enum: [embedding]
        embedding:
          type: array
          items:
            type: number
    EmbeddingUsage:
      type: object
      required:
        - prompt_tokens
        - total_tokens
      properties:
        prompt_tokens:
          type: integer
        total_tokens:
          type: integer
    CreateResponseRequest:
      type: object
      required:
        - model
        - input
      properties:
        model:
          type: string
        input:
          type: string
        tools:
          type: array
          items:
            $ref: '#/components/schemas/ResponseTool'
        text:
          $ref: '#/components/schemas/ResponseTextConfig'
    ResponseTool:
      type: object
      required:
        - type
        - name
      properties:
        type:
          type: string
          enum: [function]
        name:
          type: string
        description:
          type: string
        parameters: {}
    ResponseTextConfig:
      type: object
      properties:
        format:
          $ref: '#/components/schemas/ResponseTextFormat'
...
    CreateResponseResponse:
      type: object
      required:
        - id
        - object
        - status
        - model
        - output
        - usage
      properties:
        id:
          type: string
        object:
          type: string
          enum: [response]
        created_at:
          type: integer
          format: int64
        status:
          type: string
          enum: [completed, failed, cancelled, queued, in_progress]
        model:
          type: string
        output:
          type: array
          items:
            $ref: '#/components/schemas/ResponseOutputItem'
        usage:
          $ref: '#/components/schemas/ResponseUsage'
    ResponseOutputItem:
      type: object
      required:
        - type
      properties:
        type:
          type: string
          enum: [message, function_call]
        id:
          type: string
        call_id:
          type: string
        name:
          type: string
        arguments:
          type: string
        role:
          type: string
        content:
          type: array
          items:
            $ref: '#/components/schemas/ResponseOutputContent'
        status:
          type: string
...
    ResponseUsage:
      type: object
      required:
        - input_tokens
        - output_tokens
        - total_tokens
      properties:
        input_tokens:
          type: integer
        output_tokens:
          type: integer
        total_tokens:
          type: integer